+++ src/main/java/sixgaezzang/sidepeek/file/controller/FileController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import sixgaezzang.sidepeek.file.dto.FileUploadResponse;
+import sixgaezzang.sidepeek.file.dto.response.FileUploadResponse;
 import sixgaezzang.sidepeek.file.service.FileService;
 
 @RequestMapping("/files")
