+++ src/main/java/sixgaezzang/sidepeek/file/service/FileService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import org.springframework.util.unit.DataSize;
 import org.springframework.web.multipart.MultipartFile;
-import sixgaezzang.sidepeek.file.dto.FileUploadResponse;
+import sixgaezzang.sidepeek.file.dto.response.FileUploadResponse;
 import sixgaezzang.sidepeek.file.util.S3Properties;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
